backend/requirement.txt
django
djangorestframework
djangorestframework-simplejwt
pandas
django-cors-headers
django-debug-toolbar
Pillow
gunicorn
mysqlclient